string.c (rb_str_ascii_casemap): fix compile error.
error: implicit conversion loses integer precision: 'long' to 'int' [-Werror,-Wshorten-64-to-32]

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@55332 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ string.c
@@ -5837,7 +5837,7 @@ static void
 rb_str_ascii_casemap(VALUE source, OnigCaseFoldType *flags, rb_encoding *enc)
 {
     OnigUChar *source_current, *source_end;
-    int old_length = RSTRING_LEN(source);
+    long old_length = RSTRING_LEN(source);
     int length_or_invalid;
 
     if (old_length == 0) return;
